+++ DarbotsFtcLib/src/main/java/org/darbots/darbotsftclib/libcore/templates/servo_related/ServoType.java
@@ -11,8 +11,8 @@ public abstract class ServoType implements MotorType {
     public double getCountsPerRev(){
         return Math.abs(getMaxDeg() - getMinDeg());
     }
-    public abstract int getMaxDeg();
-    public abstract int getMinDeg();
+    public abstract double getMaxDeg();
+    public abstract double getMinDeg();
     public double getDegFromPosition(double Position){
         return Position * getCountsPerRev() + getMinDeg();
     }
